Improve local mode
In this commit, local mode is made to work on workstations
with less than 3 cores. Additionally, the need for a
separate working directory is removed, local simulations
are simply run in the current working directory. This
avoids unnecessary copying of datasets and executables.
Finally, results are saved in the local staging directory,
and on completion are organized to the results directory.