Fix killing of remaining workers
The worker threads that are still running after the main
loop in the master thread has ended were killed after a
blocking version of comm.recv(). This way, the order in
which the workers are killed was determined by the rank,
instead of by the order in which they finish. By using
the non-blocking Iprobe method, we can kill the workers
as they finish.